Move async error handling into `async-errors.js`
